Hidden Statusbar Pulldown (2/2
Currently this gives the options to turn it on/off and configure the timeout
for the Statusbar before hiding again. This is dependent on Hide Statusbar
being on since it doesnt make sense to have on without it. Thats also why I
made the changes to the mStatusBarHide part so it turns off when Hide Statusbar
is unchecked. If its written badly or whatever, please change however needed
since I suck at java.
Change-Id: I9ddd2d1a41da42eb264494872c5928aad8397933